Fix Internal.inMilliseconds() where the offset was a byte instead of an integer, preventing it from functioning properly across columns with more than 127 bytes in their qualifier.

+++ src/core/Internal.java
@@ -525,7 +525,7 @@ public final class Internal {
    * @since 2.0
    */
   public static boolean inMilliseconds(final byte[] qualifier, 
-      final byte offset) {
+      final int offset) {
     return inMilliseconds(qualifier[offset]);
   }
